Fix some warnings
Summary:
On GCC, my effective `CATCH_ERRORS` config is:

```
add_definitions("-Werror -Wall -Wextra -Wno-unknown-pragmas -Wno-unused-function -Wno-cast-align -Wno-shadow -Wno-deprecated -Wno-old-style-cast -Wno-unused-parameter -Wno-padded -Wno-missing-noreturn -Wno-sign-conversion -Wno-vla -Wno-zero-as-null-pointer-constant -Wno-switch-enum -Wno-unused-result -Wno-unused-value")
```

- `-Wno-unused-result` because of the calls to the external program `populatemailbox.sh` and such
- `-Wno-unused-value` because apparently the call to the `VERIFYEXEC_RET` macro in `imapmailsynctest.cpp` in the `createMessage` function has a statement that does nothing

Reviewers: cmollekopf

Tags: #sink

Differential Revision: https://phabricator.kde.org/D15057

+++ common/storage/entitystore.cpp
@@ -661,7 +661,7 @@ bool EntityStore::exists(const QByteArray &type, const QByteArray &uid)
     return true;
 }
 
-void EntityStore::readRevisions(const QByteArray &type, const QByteArray &uid, qint64 startingRevision,
+void EntityStore::readRevisions(const QByteArray &type, const QByteArray &uid, size_t startingRevision,
     const std::function<void(const QByteArray &uid, qint64 revision, const EntityBuffer &entity)> callback)
 {
     Q_ASSERT(d);
